DIZ for SEXPOTS.

+++ sbbs3/getmail.c
@@ -2,7 +2,7 @@
 
 /* Synchronet DLL-exported mail-related routines */
 
-/* $Id: getmail.c,v 1.8 2007/07/10 23:50:31 deuce Exp $ */
+/* $Id: getmail.c,v 1.7 2006/03/16 22:35:54 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -120,7 +120,7 @@ void DLLCALL delfattach(scfg_t* cfg, smbmsg_t* msg)
 /* of pointers to mail_t (message numbers and attributes)                   */
 /* smb_open(&smb) must be called prior										*/
 /****************************************************************************/
-mail_t* DLLCALL loadmail(smb_t* smb, int32_t* msgs, uint usernumber
+mail_t* DLLCALL loadmail(smb_t* smb, long* msgs, uint usernumber
 			   ,int which, long mode)
 {
 	ulong		l=0;
